backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
asyncpg==0.30.0
redis==5.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
langgraph==0.2.50
langchain==0.3.7
langchain-ollama==0.2.1
python-socketio==5.11.4
torch==2.5.1
numpy==1.26.4
